+++ Genesis/src/utopia/genesis/shape/shape2D/Bounds.scala
@@ -107,7 +107,7 @@ case class Bounds(position: Point, size: Size) extends Rectangular with ValueCon
     /**
      * An awt counterpart of these bounds
      */
-    def toAwt = new java.awt.Rectangle(position.x.ceil.toInt, position.y.ceil.toInt, width.ceil.toInt, height.ceil.toInt)
+    def toAwt = new java.awt.Rectangle(position.x.round.toInt, position.y.round.toInt, width.round.toInt, height.round.toInt)
     
     /**
      * The diagonal line for this rectangle. Starts at the position coordinates and goes all the